+++ history.md
@@ -1,6 +1,8 @@
 # 2022-08-01
 
-- Use custom popup menu, #3862
+- Use custom popup menu, #3862.
+- Use "first" instead of "none" for configuration `suggest.selection`.
+- Make "first" default for `suggest.selection`, like VSCode.
 
 # 2022-06-14